+++ packages/apidom-reference/test/parse/parsers/asyncapi-json-3/index.ts
@@ -1,0 +1,199 @@
+import fs from 'node:fs';
+import path from 'node:path';
+import { assert } from 'chai';
+import { NumberElement, isParseResultElement, hasElementSourceMap } from '@swagger-api/apidom-core';
+import { mediaTypes } from '@swagger-api/apidom-parser-adapter-asyncapi-json-3';
+import { fileURLToPath } from 'node:url';
+
+import File from '../../../../src/File.ts';
+import AsyncAPIJSON3Parser from '../../../../src/parse/parsers/asyncapi-json-3/index.ts';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+describe('parsers', function () {
+  context('AsyncAPIJSON3Parser', function () {
+    context('canParse', function () {
+      context('given file with .json extension', function () {
+        context('and with proper media type', function () {
+          specify('should return true', async function () {
+            const file1 = new File({
+              uri: '/path/to/asyncapi.json',
+              mediaType: mediaTypes.latest('json'),
+            });
+            const file2 = new File({
+              uri: '/path/to/asyncapi.json',
+              mediaType: mediaTypes.latest(),
+            });
+            const parser = new AsyncAPIJSON3Parser();
+
+            assert.isTrue(await parser.canParse(file1));
+            assert.isTrue(await parser.canParse(file2));
+          });
+        });
+
+        context('and with improper media type', function () {
+          specify('should return false', async function () {
+            const file = new File({
+              uri: '/path/to/asyncapi.json',
+              mediaType: 'application/vnd.oai.openapi+json;version=3.1.0',
+            });
+            const parser = new AsyncAPIJSON3Parser();
+
+            assert.isFalse(await parser.canParse(file));
+          });
+        });
+      });
+
+      context('given file with unknown extension', function () {
+        specify('should return false', async function () {
+          const file = new File({
+            uri: '/path/to/asyncapi.yaml',
+            mediaType: mediaTypes.latest(),
+          });
+          const parser = new AsyncAPIJSON3Parser({ fileExtensions: ['.json'] });
+
+          assert.isFalse(await parser.canParse(file));
+        });
+      });
+
+      context('given file with no extension', function () {
+        specify('should return false', async function () {
+          const file = new File({
+            uri: '/path/to/asyncapi',
+            mediaType: mediaTypes.latest(),
+          });
+          const parser = new AsyncAPIJSON3Parser({ fileExtensions: ['.json'] });
+
+          assert.isFalse(await parser.canParse(file));
+        });
+      });
+
+      context('given file with supported extension', function () {
+        context('and file data is buffer and can be detected as AsyncAPI 3.0.0', function () {
+          specify('should return true', async function () {
+            const url = path.join(__dirname, 'fixtures', 'sample-api.json');
+            const file = new File({
+              uri: '/path/to/async-api.json',
+              data: fs.readFileSync(url),
+            });
+            const parser = new AsyncAPIJSON3Parser();
+
+            assert.isTrue(await parser.canParse(file));
+          });
+        });
+
+        context('and file data is string and can be detected as AsyncAPI 3.0.0', function () {
+          specify('should return true', async function () {
+            const url = path.join(__dirname, 'fixtures', 'sample-api.json');
+            const file = new File({
+              uri: '/path/to/async-api.json',
+              data: fs.readFileSync(url).toString(),
+            });
+            const parser = new AsyncAPIJSON3Parser();
+
+            assert.isTrue(await parser.canParse(file));
+          });
+        });
+      });
+    });
+
+    context('parse', function () {
+      context('given AsyncApi 3.0.0 JSON data', function () {
+        specify('should return parse result', async function () {
+          const uri = path.join(__dirname, 'fixtures', 'sample-api.json');
+          const data = fs.readFileSync(uri).toString();
+          const file = new File({
+            uri,
+            data,
+            mediaType: mediaTypes.latest('json'),
+          });
+          const parser = new AsyncAPIJSON3Parser();
+          const parseResult = await parser.parse(file);
+
+          assert.isTrue(isParseResultElement(parseResult));
+        });
+      });
+
+      context('given AsyncApi 3.0.0 JSON data as buffer', function () {
+        specify('should return parse result', async function () {
+          const uri = path.join(__dirname, 'fixtures', 'sample-api.json');
+          const data = fs.readFileSync(uri);
+          const file = new File({
+            uri,
+            data,
+            mediaType: mediaTypes.latest('json'),
+          });
+          const parser = new AsyncAPIJSON3Parser();
+          const parseResult = await parser.parse(file);
+
+          assert.isTrue(isParseResultElement(parseResult));
+        });
+      });
+
+      context('given data that is not an AsyncApi 3.0.0 JSON data', function () {
+        specify('should coerce to string and parse', async function () {
+          const file = new File({
+            uri: '/path/to/file.json',
+            data: 1 as any,
+            mediaType: mediaTypes.latest(),
+          });
+          const parser = new AsyncAPIJSON3Parser();
+          const parseResult = await parser.parse(file);
+          const numberElement: NumberElement = parseResult.get(0);
+
+          assert.isTrue(isParseResultElement(parseResult));
+          assert.isTrue(numberElement.equals(1));
+        });
+      });
+
+      context('given empty file', function () {
+        specify('should return empty parse result', async function () {
+          const file = new File({
+            uri: '/path/to/file.json',
+            data: '',
+            mediaType: mediaTypes.latest(),
+          });
+          const parser = new AsyncAPIJSON3Parser();
+          const parseResult = await parser.parse(file);
+
+          assert.isTrue(isParseResultElement(parseResult));
+          assert.isTrue(parseResult.isEmpty);
+        });
+      });
+
+      context('sourceMap', function () {
+        context('given sourceMap enabled', function () {
+          specify('should decorate ApiDOM with source maps', async function () {
+            const uri = path.join(__dirname, 'fixtures', 'sample-api.json');
+            const data = fs.readFileSync(uri).toString();
+            const file = new File({
+              uri,
+              data,
+              mediaType: mediaTypes.latest(),
+            });
+            const parser = new AsyncAPIJSON3Parser({ sourceMap: true });
+            const parseResult = await parser.parse(file);
+
+            assert.isTrue(hasElementSourceMap(parseResult.api));
+          });
+        });
+
+        context('given sourceMap disabled', function () {
+          specify('should not decorate ApiDOM with source maps', async function () {
+            const uri = path.join(__dirname, 'fixtures', 'sample-api.json');
+            const data = fs.readFileSync(uri).toString();
+            const file = new File({
+              uri,
+              data,
+              mediaType: mediaTypes.latest(),
+            });
+            const parser = new AsyncAPIJSON3Parser({ sourceMap: false });
+            const parseResult = await parser.parse(file);
+
+            assert.isUndefined(parseResult.meta.get('sourceMap'));
+          });
+        });
+      });
+    });
+  });
+});