+++ packages/apidom-parser-adapter-json/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0](https://github.com/swagger-api/apidom/compare/v1.0.0-rc.4...v1.0.0) (2025-12-03)
+
+**Note:** Version bump only for package @swagger-api/apidom-parser-adapter-json
+
 # [1.0.0-rc.4](https://github.com/swagger-api/apidom/compare/v1.0.0-rc.3...v1.0.0-rc.4) (2025-11-25)
 
 **Note:** Version bump only for package @swagger-api/apidom-parser-adapter-json